http: deprecated http driver for httpadapter
