+++ .travis.sh
@@ -7,9 +7,9 @@
 # golint -set_exit_status `go list ./... | grep -Ev "(stackint/asm|vendor)"`
 # golint `go list ./... | grep -Ev "(stackint/asm|vendor)"`
 
-go build ./...
-go vet ./...
-golint ./...
+go build `go list ./... | grep -Ev "(env|vendor)"`
+go vet `go list ./... | grep -Ev "(env|vendor)"`
+golint `go list ./... | grep -Ev "(env|vendor)"`
 
 GOMAXPROCS=1 CI=true ginkgo -v --race --cover --coverprofile cover.out ./...
 covermerge http/cover.out httpadapter/cover.out ingress/cover.out > cover.out
