+++ cmd/openOrder/openOrder.go
@@ -14,8 +14,8 @@ import (
 	"time"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
-	"github.com/republicprotocol/renex-ingress-api-go/http"
-	"github.com/republicprotocol/renex-ingress-api-go/httpadapter"
+	"github.com/republicprotocol/renex-ingress-go/http"
+	"github.com/republicprotocol/renex-ingress-go/httpadapter"
 	"github.com/republicprotocol/republic-go/contract"
 	"github.com/republicprotocol/republic-go/crypto"
 	"github.com/republicprotocol/republic-go/identity"
