+++ httpadapter/httpadapter_test.go
@@ -1,275 +1,145 @@
 package httpadapter_test
 
 import (
-	"crypto/rand"
-	"encoding/base64"
-	mathRand "math/rand"
+	"bytes"
+	"encoding/json"
+	"errors"
+	"net/http"
+	"net/http/httptest"
 	"sync/atomic"
-	"time"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	. "github.com/republicprotocol/renex-ingress-api-go/httpadapter"
-
-	"github.com/republicprotocol/renex-ingress-api-go/ingress"
-	"github.com/republicprotocol/republic-go/crypto"
-	"github.com/republicprotocol/republic-go/order"
+	. "github.com/republicprotocol/renex-ingress-go/httpadapter"
 )
 
-var _ = Describe("Ingress Adapter", func() {
+type weakAdapter struct {
+	numOpened   int64
+	numCanceled int64
+}
 
-	Context("when marshaling and unmarshaling order fragment mappings", func() {
+func (adapter *weakAdapter) OpenOrder(signature string, orderFragmentMapping OrderFragmentMappings) error {
+	atomic.AddInt64(&adapter.numOpened, 1)
+	return nil
+}
 
-		var ord order.Order
-		var orderFragmentMappingIn OrderFragmentMapping
-		var podHashBytes [32]byte
-		var podHash string
+func (adapter *weakAdapter) CancelOrder(signature string, orderID string) error {
+	atomic.AddInt64(&adapter.numCanceled, 1)
+	return nil
+}
 
-		BeforeEach(func() {
-			rsaKey, err := crypto.RandomRsaKey()
-			Expect(err).ShouldNot(HaveOccurred())
-			ord, err = createOrder()
-			Expect(err).ShouldNot(HaveOccurred())
-			fragments, err := ord.Split(24, 16)
-			Expect(err).ShouldNot(HaveOccurred())
+type errAdapter struct {
+}
 
-			signatureBytes := [65]byte{}
-			_, err = rand.Read(signatureBytes[:])
-			Expect(err).ShouldNot(HaveOccurred())
+func (adapter *errAdapter) OpenOrder(signature string, orderFragmentMapping OrderFragmentMappings) error {
+	return errors.New("cannot open order")
+}
 
-			podHashBytes = [32]byte{}
-			_, err = rand.Read(podHashBytes[:])
-			Expect(err).ShouldNot(HaveOccurred())
-			podHash = base64.StdEncoding.EncodeToString(podHashBytes[:])
+func (adapter *errAdapter) CancelOrder(signature string, orderID string) error {
+	return errors.New("cannot cancel order")
+}
 
-			orderFragmentMappingIn = OrderFragmentMapping{}
-			orderFragmentMappingIn[podHash] = []OrderFragment{}
-			for i, fragment := range fragments {
-				orderFragment := ingress.OrderFragment{
-					Index: int64(i),
-				}
-				orderFragment.EncryptedFragment, err = fragment.Encrypt(rsaKey.PublicKey)
-				Expect(err).ShouldNot(HaveOccurred())
-				orderFragmentMappingIn[podHash] = append(
-					orderFragmentMappingIn[podHash],
-					MarshalOrderFragment(orderFragment))
-			}
-		})
-
-		It("should return the same data after marshaling and unmarshaling well formed data", func() {
-			ordID, orderFragmentMapping, err := UnmarshalOrderFragmentMapping(orderFragmentMappingIn)
-			Expect(err).ShouldNot(HaveOccurred())
-			Expect(ordID).Should(Equal(ord.ID))
-			Expect(orderFragmentMapping).Should(HaveLen(1))
-
-			for i, fragment := range orderFragmentMapping[podHashBytes] {
-				orderFragmentIn := MarshalOrderFragment(fragment)
-				Expect(orderFragmentIn).Should(Equal(orderFragmentMappingIn[podHash][i]))
-			}
-		})
-
-		It("should return an error for malformed order fragment IDs", func() {
-			for i := range orderFragmentMappingIn[podHash] {
-				orderFragmentMappingIn[podHash][i].ID = orderFragmentMappingIn[podHash][i].ID[1:]
-			}
-			_, _, err := UnmarshalOrderFragmentMapping(orderFragmentMappingIn)
-			Expect(err).Should(HaveOccurred())
-
-			for i := range orderFragmentMappingIn[podHash] {
-				orderFragmentMappingIn[podHash][i].ID = "this is invalid"
-			}
-			_, _, err = UnmarshalOrderFragmentMapping(orderFragmentMappingIn)
-			Expect(err).Should(HaveOccurred())
-		})
-
-		It("should return an error for malformed order fragment IDs", func() {
-			for i := range orderFragmentMappingIn[podHash] {
-				orderFragmentMappingIn[podHash][i].OrderID = orderFragmentMappingIn[podHash][i].OrderID[1:]
-			}
-			_, _, err := UnmarshalOrderFragmentMapping(orderFragmentMappingIn)
-			Expect(err).Should(HaveOccurred())
-
-			for i := range orderFragmentMappingIn[podHash] {
-				orderFragmentMappingIn[podHash][i].OrderID = "this is invalid"
-			}
-			_, _, err = UnmarshalOrderFragmentMapping(orderFragmentMappingIn)
-			Expect(err).Should(HaveOccurred())
-		})
-
-		It("should return an error for malformed pod hashes", func() {
-			orderFragmentMappingIn[podHash[16:]] = orderFragmentMappingIn[podHash]
-			_, _, err := UnmarshalOrderFragmentMapping(orderFragmentMappingIn)
-			Expect(err).Should(HaveOccurred())
-
-			delete(orderFragmentMappingIn, podHash[16:])
-			orderFragmentMappingIn["this is invalid"] = orderFragmentMappingIn[podHash]
-			_, _, err = UnmarshalOrderFragmentMapping(orderFragmentMappingIn)
-			Expect(err).Should(HaveOccurred())
-		})
-
-		It("should return an error for malformed tokens", func() {
-			for i := range orderFragmentMappingIn[podHash] {
-				orderFragmentMappingIn[podHash][i].Tokens = "this is invalid"
-			}
-			_, _, err := UnmarshalOrderFragmentMapping(orderFragmentMappingIn)
-			Expect(err).Should(HaveOccurred())
-		})
-
-		It("should return an error for malformed prices", func() {
-			for i := range orderFragmentMappingIn[podHash] {
-				orderFragmentMappingIn[podHash][i].Price = []string{}
-			}
-			_, _, err := UnmarshalOrderFragmentMapping(orderFragmentMappingIn)
-			Expect(err).Should(HaveOccurred())
-
-			for i := range orderFragmentMappingIn[podHash] {
-				orderFragmentMappingIn[podHash][i].Price = []string{"this is invalid", "this is also invalid"}
-			}
-			_, _, err = UnmarshalOrderFragmentMapping(orderFragmentMappingIn)
-			Expect(err).Should(HaveOccurred())
-		})
-
-		It("should return an error for malformed volumes", func() {
-			for i := range orderFragmentMappingIn[podHash] {
-				orderFragmentMappingIn[podHash][i].Volume = []string{}
-			}
-			_, _, err := UnmarshalOrderFragmentMapping(orderFragmentMappingIn)
-			Expect(err).Should(HaveOccurred())
-
-			for i := range orderFragmentMappingIn[podHash] {
-				orderFragmentMappingIn[podHash][i].Volume = []string{"this is invalid", "this is also invalid"}
-			}
-			_, _, err = UnmarshalOrderFragmentMapping(orderFragmentMappingIn)
-			Expect(err).Should(HaveOccurred())
-		})
-
-		It("should return an error for malformed minimum volumes", func() {
-			for i := range orderFragmentMappingIn[podHash] {
-				orderFragmentMappingIn[podHash][i].MinimumVolume = []string{}
-			}
-			_, _, err := UnmarshalOrderFragmentMapping(orderFragmentMappingIn)
-			Expect(err).Should(HaveOccurred())
-
-			for i := range orderFragmentMappingIn[podHash] {
-				orderFragmentMappingIn[podHash][i].MinimumVolume = []string{"this is invalid", "this is also invalid"}
-			}
-			_, _, err = UnmarshalOrderFragmentMapping(orderFragmentMappingIn)
-			Expect(err).Should(HaveOccurred())
-		})
-	})
+var _ = Describe("HTTP handlers", func() {
 
 	Context("when opening orders", func() {
 
-		It("should forward data to the ingress if the signature and mapping are well formed", func() {
-			ingress := &mockIngress{}
-			ingressAdapter := NewIngressAdapter(ingress)
+		It("should return status 201 for a valid request", func() {
 
-			signatureBytes := [65]byte{}
-			_, err := rand.Read(signatureBytes[:])
+			mockOrder := new(OpenOrderRequest)
+			data, err := json.Marshal(mockOrder)
 			Expect(err).ShouldNot(HaveOccurred())
-			signature := base64.StdEncoding.EncodeToString(signatureBytes[:])
 
-			orderFragmentMappingIn := OrderFragmentMapping{}
-			orderFragmentMappingIn["Td2YBy0MRYPYqqBduRmDsIhTySQUlMhPBM+wnNPWKqq="] = []OrderFragment{}
+			body := bytes.NewBuffer(data)
+			w := httptest.NewRecorder()
+			r := httptest.NewRequest("POST", "http://localhost/orders", body)
 
-			orderFragmentMappingsIn := OrderFragmentMappings{}
-			orderFragmentMappingsIn = append(orderFragmentMappingsIn, orderFragmentMappingIn)
-			err = ingressAdapter.OpenOrder(signature, orderFragmentMappingsIn)
-			Expect(err).ShouldNot(HaveOccurred())
-			Expect(atomic.LoadInt64(&ingress.numOpened)).To(Equal(int64(1)))
+			adapter := weakAdapter{}
+			server := NewServer(&adapter, &adapter)
+			server.ServeHTTP(w, r)
+
+			Expect(w.Code).To(Equal(http.StatusCreated))
+			Expect(atomic.LoadInt64(&adapter.numOpened)).To(Equal(int64(1)))
 		})
 
-		It("should not call ingress.OpenOrder if signature is invalid", func() {
-			ingress := &mockIngress{}
-			ingressAdapter := NewIngressAdapter(ingress)
-			signatureBytes := []byte{}
-			copy(signatureBytes[:], "incorrect signature")
-			orderFragmentMappingIn := OrderFragmentMapping{}
-			orderFragmentMappingIn["Td2YBy0MRYPYqqBduRmDsIhTySQUlMhPBM+wnNPWKqq="] = []OrderFragment{}
+		It("should return status 400 for an invalid request", func() {
 
-			orderFragmentMappingsIn := OrderFragmentMappings{}
-			orderFragmentMappingsIn = append(orderFragmentMappingsIn, orderFragmentMappingIn)
+			mockOrder := ""
+			data, err := json.Marshal(mockOrder)
+			Expect(err).ShouldNot(HaveOccurred())
 
-			err := ingressAdapter.OpenOrder(string(signatureBytes), orderFragmentMappingsIn)
-			Expect(err).Should(HaveOccurred())
-			Expect(atomic.LoadInt64(&ingress.numOpened)).To(Equal(int64(0)))
+			body := bytes.NewBuffer(data)
+			w := httptest.NewRecorder()
+			r := httptest.NewRequest("POST", "http://localhost/orders", body)
+
+			adapter := weakAdapter{}
+			server := NewServer(&adapter, &adapter)
+			server.ServeHTTP(w, r)
+
+			Expect(w.Code).To(Equal(http.StatusBadRequest))
+			Expect(atomic.LoadInt64(&adapter.numOpened)).To(Equal(int64(0)))
 		})
 
-		It("should not call ingress.OpenOrder if pool hash is invalid", func() {
-			ingress := &mockIngress{}
-			ingressAdapter := NewIngressAdapter(ingress)
-			signatureBytes := [65]byte{}
-			_, err := rand.Read(signatureBytes[:])
+		It("should return status 500 for adapter errors", func() {
+
+			mockOrder := new(OpenOrderRequest)
+			data, err := json.Marshal(mockOrder)
 			Expect(err).ShouldNot(HaveOccurred())
-			signature := base64.StdEncoding.EncodeToString(signatureBytes[:])
-			orderFragmentMappingIn := OrderFragmentMapping{}
-			orderFragmentMappingIn["some invalid hash"] = []OrderFragment{OrderFragment{OrderID: "thisisanorderid"}}
 
-			orderFragmentMappingsIn := OrderFragmentMappings{}
-			orderFragmentMappingsIn = append(orderFragmentMappingsIn, orderFragmentMappingIn)
+			body := bytes.NewBuffer(data)
+			w := httptest.NewRecorder()
+			r := httptest.NewRequest("POST", "http://localhost/orders", body)
 
-			err = ingressAdapter.OpenOrder(signature, orderFragmentMappingsIn)
-			Expect(err).Should(HaveOccurred())
-			Expect(atomic.LoadInt64(&ingress.numOpened)).To(Equal(int64(0)))
+			adapter := errAdapter{}
+			server := NewServer(&adapter, &adapter)
+			server.ServeHTTP(w, r)
+
+			Expect(w.Code).To(Equal(http.StatusInternalServerError))
 		})
 	})
 
 	Context("when canceling orders", func() {
 
-		It("should call ingress.CancelOrder if signature and orderID is valid", func() {
-			ingress := &mockIngress{}
-			ingressAdapter := NewIngressAdapter(ingress)
-			signatureBytes := [65]byte{}
-			_, err := rand.Read(signatureBytes[:])
-			Expect(err).ShouldNot(HaveOccurred())
-			signature := base64.StdEncoding.EncodeToString(signatureBytes[:])
-			err = ingressAdapter.CancelOrder(signature, "Td2YBy0MRYPYqqBduRmDsIhTySQUlMhPBM+wnNPWKqq=")
+		It("should return status 200 for a valid request", func() {
 
-			Expect(err).To(BeNil())
-			Expect(atomic.LoadInt64(&ingress.numCanceled)).To(Equal(int64(1)))
+			orderID := "vrZhWU3VV9LRIriRvuzT9CbVc57wQhbQ"
+			signature := "Td2YBy0MRYPYqqBduRmDsIhTySQUlMhPBM+wnNPWKqq="
+			w := httptest.NewRecorder()
+			r := httptest.NewRequest("DELETE", "http://localhost/orders?id="+orderID+"&signature="+signature, nil)
+
+			adapter := weakAdapter{}
+			server := NewServer(&adapter, &adapter)
+			server.ServeHTTP(w, r)
+
+			Expect(w.Code).To(Equal(http.StatusOK))
+			Expect(atomic.LoadInt64(&adapter.numCanceled)).To(Equal(int64(1)))
 		})
 
-		It("should not call ingress.CancelOrder if orderID is invalid", func() {
-			ingress := &mockIngress{}
-			ingressAdapter := NewIngressAdapter(ingress)
-			signatureBytes := [65]byte{}
-			_, err := rand.Read(signatureBytes[:])
-			Expect(err).ShouldNot(HaveOccurred())
-			signature := base64.StdEncoding.EncodeToString(signatureBytes[:])
-			err = ingressAdapter.CancelOrder(signature, "")
-			Expect(err).Should(Equal(ErrInvalidOrderIDLength))
-			Expect(atomic.LoadInt64(&ingress.numCanceled)).To(Equal(int64(0)))
+		It("should return status 400 for an invalid signature", func() {
+
+			orderID := "vrZhWU3VV9LRIriRvuzT9CbVc57wQhbQ"
+			w := httptest.NewRecorder()
+			r := httptest.NewRequest("DELETE", "http://localhost/orders?id="+orderID, nil)
+
+			adapter := weakAdapter{}
+			server := NewServer(&adapter, &adapter)
+			server.ServeHTTP(w, r)
+
+			Expect(w.Code).To(Equal(http.StatusBadRequest))
+			Expect(atomic.LoadInt64(&adapter.numOpened)).To(Equal(int64(0)))
+		})
+
+		It("should return status 500 for adapter errors", func() {
+
+			orderID := "vrZhWU3VV9LRIriRvuzT9CbVc57wQhbQ"
+			signature := "Td2YBy0MRYPYqqBduRmDsIhTySQUlMhPBM+wnNPWKqq="
+			w := httptest.NewRecorder()
+			r := httptest.NewRequest("DELETE", "http://localhost/orders?id="+orderID+"&signature="+signature, nil)
+
+			adapter := errAdapter{}
+			server := NewServer(&adapter, &adapter)
+			server.ServeHTTP(w, r)
+
+			Expect(w.Code).To(Equal(http.StatusInternalServerError))
 		})
 	})
 })
-
-type mockIngress struct {
-	numOpened   int64
-	numCanceled int64
-}
-
-func (ingress *mockIngress) Sync(done <-chan struct{}) <-chan error {
-	return nil
-}
-
-func (ingress *mockIngress) OpenOrder(signature [65]byte, orderID order.ID, orderFragmentMapping ingress.OrderFragmentMappings) error {
-	atomic.AddInt64(&ingress.numOpened, 1)
-	return nil
-}
-
-func (ingress *mockIngress) CancelOrder(signature [65]byte, orderID order.ID) error {
-	atomic.AddInt64(&ingress.numCanceled, 1)
-	return nil
-}
-
-func (ingress *mockIngress) ProcessRequests(done <-chan struct{}) <-chan error {
-	return nil
-}
-
-func createOrder() (order.Order, error) {
-	parity := order.ParityBuy
-	price := uint64(mathRand.Intn(2000))
-	volume := uint64(mathRand.Intn(2000))
-	nonce := uint64(mathRand.Intn(1000000000))
-	return order.NewOrder(order.TypeLimit, parity, order.SettlementRenEx, time.Now().Add(time.Hour), order.TokensETHREN, order.NewCoExp(price, 26), order.NewCoExp(volume, 26), order.NewCoExp(volume, 26), nonce), nil
-}
