+++ httpadapter/marshal_test.go
@@ -1,0 +1,1 @@
+package httpadapter_test
