+++ ingress/ingress_test.go
@@ -12,7 +12,7 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	. "github.com/republicprotocol/renex-ingress-api-go/ingress"
+	. "github.com/republicprotocol/renex-ingress-go/ingress"
 
 	"github.com/republicprotocol/republic-go/crypto"
 	"github.com/republicprotocol/republic-go/identity"
